backend/requirements/base.txt
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.13.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
pillow==12.2.0
requests==2.31.0
resend==2.1.0
google-auth==2.49.2
google-genai==1.73.1